requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
drf-yasg==1.20.0
Pillow==8.4.0
djangorestframework==3.12.4
